Unify how we create random inputs for auto-tuning
Summary: We're creating autotune inputs slightly differently when autotuning in-process vs. in a subprocess: One implementation is in TensorMeta.to_tensor() and another in AlgorithmSelectorCache.benchmark_example_value. Move the TensorMeta definition to select_algorith.py and call that implementation from AlgorithmSelectorCache.benchmark_example_value().

Test Plan: Existing unit tests

[ghstack-poisoned]
